+++ py/test/test_database.py
@@ -82,7 +82,7 @@ class DatabaseTests(unittest.TestCase):
             db.add_tag(items[0], 'food / groceries')
             db.add_tag(items[1], 'food / groceries')
             filter = BusinessFilter(db)
-            self.assertEqual(filter('business'), receipts)
+            self.assertEqual(filter(name='business'), receipts)
             filter = ItemFilter(db)
             self.assertEqual(filter(name='test integer'), items[1:2])
             filter = PriceFilter(db)
@@ -94,7 +94,7 @@ class DatabaseTests(unittest.TestCase):
             self.assertEqual(filter(path='food / groceries'), items[0:2])
             self.assertEqual(filter(path='food'), items[0:2])
             filter = ItemTags(db)
-            self.assertEqual(tuple(filter(items[1]))[0], 'food / groceries')
+            self.assertEqual(tuple(filter(item=items[1]))[0], 'food / groceries')
             filter = DateFilter(db)
             self.assertEqual(filter(lower_date_boundary=DatabaseTests.yesterday), items)
             self.assertEqual(filter(upper_date_boundary=DatabaseTests.tomorrow), items)
